# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "B&B AI — API",
    "description": "Agente contable IA enterprise para despachos contables. Endpoints /api/v1/* requieren API key en el header X-API-Key.",
    "contact": {
      "name": "B&B AI",
      "email": "ventas@b2b-ai.local"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/reconciliation/upload": {
      "post": {
        "tags": [
          "reconciliation"
        ],
        "summary": "Sube un estado de cuenta (CSV/PDF) y lo carga.",
        "description": "Sube el estado de cuenta como multipart (campo `file`) e indica el\nbanco (campo `bank`): bbva, banorte, santander, hsbc o generico.\nParsea los movimientos y los carga en la sesión del tenant.",
        "operationId": "upload_statement_api_v1_reconciliation_upload_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_statement_api_v1_reconciliation_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/v1/reconciliation/matches": {
      "get": {
        "tags": [
          "reconciliation"
        ],
        "summary": "Cruces automáticos calculados (con confidence).",
        "description": "Recalcula (si `refresh=True`) y devuelve los cruces con su\nconfidence score 0-100. Cruza contra las facturas del tenant en DB.",
        "operationId": "matches_api_v1_reconciliation_matches_get",
        "parameters": [
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Refresh"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/v1/reconciliation/confirm": {
      "post": {
        "tags": [
          "reconciliation"
        ],
        "summary": "Confirma manualmente un cruce transacción-factura.",
        "description": "Confirma que la transacción `transaction_id` concilia la factura\n`invoice_id`. El cruce queda marcado method=manual, confidence=100.",
        "operationId": "confirm_api_v1_reconciliation_confirm_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/v1/reconciliation/report": {
      "get": {
        "tags": [
          "reconciliation"
        ],
        "summary": "Reporte de conciliación de la sesión actual.",
        "description": "Genera el reporte agregado de conciliación (conciliados,\npendientes, tasas, por método) para el tenant.",
        "operationId": "report_api_v1_reconciliation_report_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_statement_api_v1_reconciliation_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "bank": {
            "type": "string",
            "title": "Bank",
            "default": "generico"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_statement_api_v1_reconciliation_upload_post"
      },
...
      "ConfirmRequest": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "invoice_id": {
            "type": "string",
            "title": "Invoice Id"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "invoice_id"
        ],
        "title": "ConfirmRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key del tenant (tabla api_keys) o la key de servicio (env B2B_API_KEY). Header X-API-Key."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Token del Portal del cliente (POST /portal/auth/login)."
      },
      "JWTAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT access token de auth enterprise (POST /api/v1/auth/login)."
      }
    }
  },
  "servers": [
    {
      "url": "https://api.b2b-ai.local",
      "description": "Producción"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local (uvicorn)"
    }
  ]
}